refactor: use strum discriminants to generate the code for DataTypeName
