+++ src/common/Cargo.toml
@@ -41,6 +41,8 @@ serde_derive = "1"
 serde_json = "1"
 smallvec = "1"
 spin = "0.9"
+strum = "0.24"
+strum_macros = "0.24"
 thiserror = "1"
 tokio = { version = "0.2", package = "madsim-tokio", features = ["rt", "rt-multi-thread", "sync", "macros", "time", "signal"] }
 tokio-stream = "0.1"
