+++ src/tests/sqlsmith/Cargo.toml
@@ -13,6 +13,7 @@ lazy_static = "1"
 libtest-mimic = "0.5"
 paste = "1"
 rand = { version = "0.8", features = ["small_rng"] }
+risingwave_common = { path = "../../common" }
 risingwave_expr = { path = "../../expr" }
 risingwave_frontend = { path = "../../frontend" }
 risingwave_sqlparser = { path = "../../sqlparser" }
