+++ src/tests/sqlsmith/src/lib.rs
@@ -17,8 +17,8 @@ use std::vec;
 use itertools::Itertools;
 use rand::prelude::SliceRandom;
 use rand::Rng;
+use risingwave_common::types::DataTypeName;
 use risingwave_frontend::bind_data_type;
-use risingwave_frontend::expr::DataTypeName;
 use risingwave_sqlparser::ast::{
     BinaryOperator, ColumnDef, Cte, Expr, Ident, Join, JoinConstraint, JoinOperator, ObjectName,
     OrderByExpr, Query, Select, SelectItem, SetExpr, Statement, TableWithJoins, With,
