+++ src/tests/sqlsmith/src/utils.rs
@@ -15,7 +15,7 @@
 use std::mem;
 
 use rand::Rng;
-use risingwave_frontend::expr::DataTypeName;
+use risingwave_common::types::DataTypeName;
 use risingwave_sqlparser::ast::{
     DataType, FunctionArg, FunctionArgExpr, TableAlias, TableFactor, TableWithJoins,
 };
